refactor: rename ChatComposer blocks, add dropdown menus, move FollowUpQuestions
- Rename ChatComposerComposerWithError → ChatComposerValidation
  Labels: 'Error message (with top position)' / 'Warning message (with bottom position)'
- Rename ChatComposerSimpleComposer → ChatComposerSimple
  Single composer with default placeholder, 16:9 aspect ratio
- Rename ChatComposerFullFeaturedComposer → ChatComposerFullFeatured
  Footer actions now use XDSDropdownMenu (Auto + Settings), drawer is collapsible
- Rename ChatComposerComposerWithAttachments → ChatComposerAttachments
  Single example with collapsible drawer (6 files)
- Update ChatComposerFooterActions to use XDSDropdownMenu for Auto and Settings
  (matches the AI Chat page template pattern)
- Update ChatComposerStreaming to 16:9 aspect ratio
- Move ChatComposerFollowUpQuestions → ChatComposerDrawer directory
  Renamed to ChatComposerDrawerFollowUpQuestions, fixed raw SVG icons
  to use heroicons + XDSIcon
