+++ packages/cli/templates/blocks/components/ChatComposer/ChatComposerAttachments.doc.mjs
@@ -1,0 +1,9 @@
+/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
+export const doc = {
+  type: 'block',
+  name: 'ChatComposer — Attachments',
+  description: 'Chat composer with removable file tokens in a collapsible drawer. Use when users can attach files or context to their message.',
+  isReady: true,
+  aspectRatio: 4 / 3,
+  componentsUsed: ['ChatComposer', 'Token', 'Layout'],
+};
