--- packages/cli/templates/blocks/components/ChatComposer/ChatComposerComposerWithAttachments.doc.mjs
@@ -1,9 +1,0 @@
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  name: 'ChatComposer — Attachments',
-  description: 'Chat composer with removable file tokens in a collapsible drawer and a context progress bar. Use when users can attach files or context to their message.',
-  isReady: true,
-  aspectRatio: 4 / 3,
-  componentsUsed: ['ChatComposer', 'Token', 'ProgressBar', 'Button', 'Icon', 'Layout', 'Text'],
-};
