--- packages/cli/templates/blocks/components/ChatComposer/ChatComposerComposerWithError.doc.mjs
@@ -1,9 +1,0 @@
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  name: 'ChatComposer — Status',
-  description: 'Chat composer with error and warning status messages. Status can appear above or below the composer to surface validation or system feedback.',
-  isReady: true,
-  aspectRatio: 4 / 3,
-  componentsUsed: ['ChatComposer', 'Layout', 'Text'],
-};
