--- packages/cli/templates/blocks/components/ChatComposer/ChatComposerFollowUpQuestions.doc.mjs
@@ -1,10 +1,0 @@
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  name: 'ChatComposer — Follow-Up Questions',
-  description:
-    'Chat composer with a collapsible drawer containing navigable follow-up questions and selectable lettered options.',
-  isReady: true,
-  aspectRatio: 4 / 3,
-  componentsUsed: ['ChatComposer', 'List', 'Badge', 'Text', 'Button'],
-};
