+++ packages/cli/templates/blocks/components/ChatComposer/ChatComposerFooterActions.doc.mjs
@@ -2,8 +2,8 @@
 export const doc = {
   type: 'block',
   name: 'ChatComposer — Footer Actions',
-  description: 'Chat composer with footer action buttons like a model selector and a mic button in the send actions slot. Use for AI chat interfaces that let users pick a model or mode.',
+  description: 'Chat composer with dropdown menus for a model selector and settings in the footer, and a mic button in the send actions slot.',
   isReady: true,
   aspectRatio: 4 / 3,
-  componentsUsed: ['ChatComposer', 'Button', 'Icon', 'Layout', 'Text'],
+  componentsUsed: ['ChatComposer', 'Button', 'DropdownMenu', 'Icon', 'Layout', 'Text'],
 };
