+++ packages/cli/templates/blocks/components/ChatComposer/ChatComposerFullFeatured.doc.mjs
@@ -1,0 +1,9 @@
+/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
+export const doc = {
+  type: 'block',
+  name: 'ChatComposer — Full Featured',
+  description: 'Chat composer with all slots populated — collapsible attachment drawer, header actions, context progress bar, footer dropdown menus, and mic button. Shows the maximum composer configuration.',
+  isReady: true,
+  aspectRatio: 4 / 3,
+  componentsUsed: ['ChatComposer', 'Token', 'Button', 'DropdownMenu', 'Icon', 'ProgressBar', 'Layout', 'Text'],
+};
