+++ packages/cli/templates/blocks/components/ChatComposer/ChatComposerFullFeatured.tsx
@@ -4,17 +4,20 @@ import {useState} from 'react';
 import {XDSChatComposer, XDSChatComposerDrawer} from '@xds/core/Chat';
 import {XDSToken} from '@xds/core/Token';
 import {XDSButton} from '@xds/core/Button';
+import {XDSDropdownMenu} from '@xds/core/DropdownMenu';
 import {XDSIcon} from '@xds/core/Icon';
 import {XDSProgressBar} from '@xds/core/ProgressBar';
 import {XDSStack} from '@xds/core/Layout';
 import {XDSText} from '@xds/core/Text';
 import {
   AtSymbolIcon,
+  Cog6ToothIcon,
   MicrophoneIcon,
   PaperClipIcon,
+  SparklesIcon,
 } from '@heroicons/react/24/outline';
 
-export default function ChatComposerFullFeaturedComposer() {
+export default function ChatComposerFullFeatured() {
   const [isStreaming, setIsStreaming] = useState(false);
 
   return (
@@ -32,9 +35,12 @@ export default function ChatComposerFullFeaturedComposer() {
         onStop={() => setIsStreaming(false)}
         placeholder="Ask me anything..."
         drawer={
-          <XDSChatComposerDrawer>
+          <XDSChatComposerDrawer count={5}>
             <XDSToken label="design-spec.pdf" onRemove={() => {}} />
             <XDSToken label="requirements.docx" onRemove={() => {}} />
+            <XDSToken label="wireframes.fig" onRemove={() => {}} />
+            <XDSToken label="api-spec.yaml" onRemove={() => {}} />
+            <XDSToken label="user-research.csv" onRemove={() => {}} />
           </XDSChatComposerDrawer>
         }
         headerActions={
@@ -60,8 +66,37 @@ export default function ChatComposerFullFeaturedComposer() {
         }
         footerActions={
           <>
-            <XDSButton label="Auto" variant="ghost" size="md" />
-            <XDSButton label="Settings" variant="ghost" size="md" />
+            <XDSDropdownMenu
+              button={{
+                label: 'Auto',
+                variant: 'ghost',
+                size: 'md',
+                icon: <XDSIcon icon={SparklesIcon} size="sm" />,
+                children: 'Auto',
+              }}
+              menuWidth={200}
+              items={[
+                {label: 'Auto', onClick: () => {}},
+                {label: 'GPT-4o', onClick: () => {}},
+                {label: 'Claude 3.5', onClick: () => {}},
+                {label: 'Llama 3', onClick: () => {}},
+              ]}
+            />
+            <XDSDropdownMenu
+              button={{
+                label: 'Settings',
+                variant: 'ghost',
+                size: 'md',
+                icon: <XDSIcon icon={Cog6ToothIcon} size="sm" />,
+                children: 'Settings',
+              }}
+              menuWidth={200}
+              items={[
+                {label: 'Preferences', onClick: () => {}},
+                {label: 'Keyboard shortcuts', onClick: () => {}},
+                {label: 'About', onClick: () => {}},
+              ]}
+            />
           </>
         }
         sendActions={
