--- packages/cli/templates/blocks/components/ChatComposer/ChatComposerFullFeaturedComposer.doc.mjs
@@ -1,9 +1,0 @@
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  name: 'ChatComposer — Full Featured',
-  description: 'Chat composer with all slots populated — attachment drawer, header actions, context progress bar, footer model selector, and mic button. Shows the maximum composer configuration.',
-  isReady: true,
-  aspectRatio: 4 / 3,
-  componentsUsed: ['ChatComposer', 'Token', 'Button', 'Icon', 'ProgressBar', 'Layout', 'Text'],
-};
