+++ packages/cli/templates/blocks/components/ChatComposer/ChatComposerSimple.doc.mjs
@@ -1,0 +1,9 @@
+/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
+export const doc = {
+  type: 'block',
+  name: 'ChatComposer — Simple',
+  description: 'Minimal chat composer with a placeholder and submit handler. The simplest way to drop a message input into a page.',
+  isReady: true,
+  aspectRatio: 16 / 9,
+  componentsUsed: ['ChatComposer', 'Layout'],
+};
