+++ packages/cli/templates/blocks/components/ChatComposer/ChatComposerSimple.tsx
@@ -1,0 +1,16 @@
+'use client';
+
+import {XDSChatComposer} from '@xds/core/Chat';
+import {XDSStack} from '@xds/core/Layout';
+
+export default function ChatComposerSimple() {
+  return (
+    <XDSStack direction="vertical" style={{width: '100%', maxWidth: 450}}>
+      <XDSChatComposer
+        onSubmit={value => {
+          console.log('Sent:', value);
+        }}
+      />
+    </XDSStack>
+  );
+}
