--- packages/cli/templates/blocks/components/ChatComposer/ChatComposerSimpleComposer.tsx
@@ -1,33 +1,0 @@
-'use client';
-
-import {XDSChatComposer} from '@xds/core/Chat';
-import {XDSStack} from '@xds/core/Layout';
-import {XDSText} from '@xds/core/Text';
-
-export default function ChatComposerSimpleComposer() {
-  return (
-    <XDSStack direction="vertical" gap={4} style={{width: '100%', maxWidth: 450}}>
-      <XDSStack direction="vertical" gap={1}>
-        <XDSText type="supporting" color="secondary">
-          Default placeholder
-        </XDSText>
-        <XDSChatComposer
-          onSubmit={value => {
-            console.log('Sent:', value);
-          }}
-        />
-      </XDSStack>
-      <XDSStack direction="vertical" gap={1}>
-        <XDSText type="supporting" color="secondary">
-          Custom placeholder
-        </XDSText>
-        <XDSChatComposer
-          onSubmit={value => {
-            console.log('Sent:', value);
-          }}
-          placeholder="Ask me anything..."
-        />
-      </XDSStack>
-    </XDSStack>
-  );
-}
