+++ packages/cli/templates/blocks/components/ChatComposer/ChatComposerValidation.doc.mjs
@@ -1,8 +1,8 @@
 /** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
 export const doc = {
   type: 'block',
-  name: 'ChatComposer — Simple',
-  description: 'Minimal chat composer with a placeholder and submit handler. The simplest way to drop a message input into a page.',
+  name: 'ChatComposer — Validation',
+  description: 'Chat composer with error and warning status messages. Status can appear above or below the composer to surface validation or system feedback.',
   isReady: true,
   aspectRatio: 4 / 3,
   componentsUsed: ['ChatComposer', 'Layout', 'Text'],
